[ricos-content] feat(comapareDraftContent) refactored from ricos-editor
