+++ packages/ricos-content/web/libs/comapareDraftContent/package.json
@@ -1,0 +1,5 @@
+{
+    "main": "../../dist/lib/comapareDraftContent.cjs.js",
+    "module": "../../dist/lib/comapareDraftContent.js",
+    "types": "../../dist/lib/comapareDraftContent.d.ts"
+}
