+++ packages/ricos-editor/web/src/utils/editorUtils.ts
@@ -5,9 +5,10 @@ import {
   convertFromRaw,
 } from 'wix-rich-content-editor/libs/editorStateConversion';
 import { EditorProps } from 'draft-js';
-import { debounce, pick, sortBy, isEqual, isEmpty } from 'lodash';
+import { debounce, pick } from 'lodash';
 import { emptyState, DRAFT_EDITOR_PROPS } from 'ricos-common';
-import { compare, isContentStateEmpty } from 'ricos-content';
+import { isContentStateEmpty } from 'ricos-content';
+import { isContentEqual } from 'ricos-content/libs/comapareDraftContent';
 import { DraftContent, isSSR } from 'wix-rich-content-common';
 import { EditorDataInstance, OnContentChangeFunction, ContentStateGetter } from '../index';
 import errorBlocksRemover from './errorBlocksRemover';
@@ -20,45 +21,6 @@ export const ONCHANGE_DEBOUNCE_TIME = 200;
 const wait = ms => {
   return new Promise(resolve => setTimeout(resolve, ms));
 };
-
-// compares ranges regardless style order, e.g. ['BOLD', 'ITALIC'] equals ['ITALIC', 'BOLD']
-function areRangesEqual(blockRanges1, blockRanges2, sortKey) {
-  return (
-    blockRanges1.length === blockRanges2.length &&
-    isEqual(
-      sortBy(blockRanges1, r => r[sortKey]),
-      sortBy(blockRanges2, r => r[sortKey])
-    )
-  );
-}
-
-function isTextAlignmentEqual(block1, block2) {
-  const isBlock1AlignedLeft = !block1.data.textAlignment || block1.data.textAlignment === 'left';
-  const isBlock2AlignedLeft = !block2.data.textAlignment || block2.data.textAlignment === 'left';
-  return (
-    isBlock1AlignedLeft === isBlock2AlignedLeft ||
-    block1.data.textAlignment === block2.data.textAlignment
-  );
-}
-
-function areBlockFieldsEqual(block1, block2) {
-  return (
-    block1.text === block2.text &&
-    block1.depth === block2.depth &&
-    block1.type === block2.type &&
-    areRangesEqual(block1.inlineStyleRanges, block2.inlineStyleRanges, 'style') &&
-    areRangesEqual(block1.entityRanges, block2.entityRanges, 'key') &&
-    isTextAlignmentEqual(block1, block2) &&
-    isEmpty(compare(block1.data, block2.data, { verbose: false, ignoredKeys: ['textAlignment'] }))
-  );
-}
-
-function areBlocksEqual(currentStateBlocks, initialStateBlocks) {
-  return (
-    currentStateBlocks.length === initialStateBlocks.length &&
-    currentStateBlocks.every((block, i) => areBlockFieldsEqual(block, initialStateBlocks[i]))
-  );
-}
 
 export function createDataConverter(
   onContentChange?: OnContentChangeFunction,
@@ -96,16 +58,12 @@ export function createDataConverter(
     lastContent: DraftContent,
     initialContent: DraftContent
   ) => {
-    const initialBlocksEqual = areBlocksEqual(currContent.blocks, initialContent.blocks);
-    const initialEntitiesEqual = isEmpty(
-      compare(currContent.entityMap, initialContent.entityMap, { verbose: false })
-    );
-    const lastBlocksEqual = areBlocksEqual(currContent.blocks, lastContent.blocks);
-    const lastEntitiesEqual = isEmpty(compare(currContent.entityMap, lastContent.entityMap));
+    const initialContentEqual = isContentEqual(currContent, initialContent);
+    const lastContentEqual = isContentEqual(currContent, lastContent);
     currTraits = {
       isEmpty: isContentStateEmpty(currContent),
-      isContentChanged: !(initialBlocksEqual && initialEntitiesEqual),
-      isLastChangeEdit: !(lastBlocksEqual && lastEntitiesEqual),
+      isContentChanged: !initialContentEqual,
+      isLastChangeEdit: !lastContentEqual,
     };
   };
